+++ modules/game/include/game_object.h
@@ -8,15 +8,15 @@
 class GameObject {
 protected:
 	bool alive;
-	double x, y;
-	double oldX, oldY;
-	double size;
-	double maxSize;
-	double speed;
-	double angle;
-	double height;
+	float x, y;
+	float oldX, oldY;
+	float size;
+	float maxSize;
+	float speed;
+	float angle;
+	float height;
 
-	double cameraOffsetX, cameraOffsetY;
+	float cameraOffsetX, cameraOffsetY;
 
 public:
 	GameObject() {
@@ -37,34 +37,34 @@ public:
 
 	GameObject *itSelf(void) {return this;}
 
-	double getX(void) {return x;}
-	double getY(void) {return y;}
-	void setXY(double newX, double newY) {x = newX; y = newY;}
-	void setSpeed(double newSpeed) {speed = newSpeed;}
-	double getSpeed(void) {return speed;}
-	double getSize(void) {return size;}
-	void setSize(double value) {size=value;}
-	double getAngle(void) {return angle;}
-	void setHeight(double h) {height=h;}
-	void setCameraOffset(double X,double Y){cameraOffsetX=X;cameraOffsetY=Y;}
+	float getX(void) {return x;}
+	float getY(void) {return y;}
+	void setXY(float newX, float newY) {x = newX; y = newY;}
+	void setSpeed(float newSpeed) {speed = newSpeed;}
+	float getSpeed(void) {return speed;}
+	float getSize(void) {return size;}
+	void setSize(float value) {size=value;}
+	float getAngle(void) {return angle;}
+	void setHeight(float h) {height=h;}
+	void setCameraOffset(float X,float Y){cameraOffsetX=X;cameraOffsetY=Y;}
 
 	void backupXY(void) {oldX = x; oldY = y;}
 	void restoreXY(void) {x = oldX; y = oldY;}
 
-	double distance(GameObject *obj) {
-		double x1 = obj->getX();
-		double y1 = obj->getY();
-		double x2 = x;
-		double y2 = y;
+	float distance(GameObject *obj) {
+		float x1 = obj->getX();
+		float y1 = obj->getY();
+		float x2 = x;
+		float y2 = y;
 
-		double d=x2-x1;
-		double d2=y2-y1;
-		double result = sqrt(pow(x - obj->getX(), 2.0) + pow(y - obj->getY(), 2.0));
+		float d=x2-x1;
+		float d2=y2-y1;
+		float result = sqrt(pow(x - obj->getX(), 2.0) + pow(y - obj->getY(), 2.0));
 
 		return result;
 	}
 
-	double getAngleAt(GameObject *obj) {
+	float getAngleAt(GameObject *obj) {
 		return(atan2(obj->y - y, obj->x - x));
 	}
 
@@ -72,5 +72,5 @@ public:
     
   }
 	
-	virtual void takeDamage(double damage) {};
+	virtual void takeDamage(float damage) {};
 };
