+++ modules/game/include/misc.h
@@ -4,18 +4,18 @@
 #include "defines.h"
 #include "gameObject.h"
 
-const double _PI = 3.1415926535897932384626;//4338327950288419716939937510582097494459230781640628620899;
-const double ATTACK_RANGE_MELEE = 5.0,
+const float _PI = 3.1415926535897932384626;//4338327950288419716939937510582097494459230781640628620899;
+const float ATTACK_RANGE_MELEE = 5.0,
 	         ATTACK_RANGE_RANGED = 200.0;
 
 // way searching
-//double lx1, ly1, lx2, ly2; // coords of intersected block line
+//float lx1, ly1, lx2, ly2; // coords of intersected block line
 
-inline bool isIntersection(double ax1, double ay1, double ax2, double ay2, double bx1, double by1, double bx2, double by2/*,double *coords,int *count*/) {
-	double v1=(bx2-bx1)*(ay1-by1)-(by2-by1)*(ax1-bx1);
-	double v2=(bx2-bx1)*(ay2-by1)-(by2-by1)*(ax2-bx1);
-	double v3=(ax2-ax1)*(by1-ay1)-(ay2-ay1)*(bx1-ax1);
-	double v4=(ax2-ax1)*(by2-ay1)-(ay2-ay1)*(bx2-ax1);
+inline bool isIntersection(float ax1, float ay1, float ax2, float ay2, float bx1, float by1, float bx2, float by2/*,float *coords,int *count*/) {
+	float v1=(bx2-bx1)*(ay1-by1)-(by2-by1)*(ax1-bx1);
+	float v2=(bx2-bx1)*(ay2-by1)-(by2-by1)*(ax2-bx1);
+	float v3=(ax2-ax1)*(by1-ay1)-(ay2-ay1)*(bx1-ax1);
+	float v4=(ax2-ax1)*(by2-ay1)-(ay2-ay1)*(bx2-ax1);
 	bool b=((v1*v2<0)&&(v3*v4<0));
 	//if(b) {
 	//	int c = *count;
@@ -29,31 +29,31 @@ inline bool isIntersection(double ax1, double ay1, double ax2, double ay2, doubl
 	//return((v1*v2<0)&&(v3*v4<0));
 }
 
-inline bool isMapContainsCoord(double x,double y) {
+inline bool isMapContainsCoord(float x,float y) {
 	return(x<RESOLUTION_X && y<RESOLUTION_Y && x>0 && y>0);
 }
 
-inline bool isSegmentsParallel(double x1, double y1, double x2, double y2, double x3, double y3, double x4, double y4) {
+inline bool isSegmentsParallel(float x1, float y1, float x2, float y2, float x3, float y3, float x4, float y4) {
 	return(((x2-x1)/(y2-y1))==((x4-x3)/(y4-y3)));
 }
 
 //angle between coords in radians
-inline double abc(double x1, double y1, double x2, double y2) {
+inline float abc(float x1, float y1, float x2, float y2) {
 	return(atan2(y2-y1, x2-x1));
 }
 
 //distance between objects
-inline double dbo(GAME_OBJECT *f, GAME_OBJECT *s) {
-	double x1=f->getX();
-	double y1=f->getY();
-	double x2=s->getX();
-	double y2=s->getY();
+inline float dbo(GAME_OBJECT *f, GAME_OBJECT *s) {
+	float x1=f->getX();
+	float y1=f->getY();
+	float x2=s->getX();
+	float y2=s->getY();
 	return sqrt(pow(x2-x1,2.0)+pow(y2-y1,2.0));
 	//return(sqrt((x2-x1)*(x2-x1)+(y2-y1)*(y2-y1)));
 }
 
 //distance between coords
-inline double dbc(double x1, double y1, double x2, double y2) {
+inline float dbc(float x1, float y1, float x2, float y2) {
 	return sqrt(pow(x2-x1,2.0)+pow(y2-y1,2.0));
 }
